split byte coding in two files (signed/unsigned)
