+++ src/coding/ibyte.rs
@@ -2,33 +2,7 @@ use super::{Decodeable, Encodeable};
 use std::collections::VecDeque;
 use std::io;
 
-pub type MinecraftUnsignedByte = u8;
 pub type MinecraftByte = i8;
-
-impl Decodeable<MinecraftUnsignedByte, io::Error> for VecDeque<u8> {
-    fn decode(&mut self) -> Result<MinecraftUnsignedByte, io::Error> {
-        let value = self.pop_front();
-
-        if value.is_some() {
-            Ok(value.unwrap())
-        } else {
-            Err(io::Error::new(
-                io::ErrorKind::InvalidInput,
-                "A minecraft byte consists of one byte, but there is none!",
-            ))
-        }
-    }
-}
-
-impl Encodeable for MinecraftUnsignedByte {
-    fn encode(&self) -> VecDeque<u8> {
-        VecDeque::from(vec![*self])
-    }
-
-    fn byte_length(&self) -> u8 {
-        1
-    }
-}
 
 impl Decodeable<MinecraftByte, io::Error> for VecDeque<u8> {
     fn decode(&mut self) -> Result<MinecraftByte, io::Error> {
@@ -58,41 +32,12 @@ impl Encodeable for MinecraftByte {
 
 #[cfg(test)]
 mod tests {
-    use super::{Decodeable, Encodeable, MinecraftUnsignedByte, MinecraftByte};
+    use super::{Decodeable, Encodeable, MinecraftByte};
     use std::collections::VecDeque;
     use std::io;
 
     #[test]
-    fn test_decoding_unsigned() {
-        let mappings: Vec<(MinecraftUnsignedByte, Vec<u8>)> =
-            vec![(0x01, vec![0x01]), (0x00, vec![0x00])];
-
-        for mapping in mappings {
-            let actual: MinecraftUnsignedByte = VecDeque::from(mapping.1).decode().unwrap();
-
-            assert_eq!(mapping.0, actual);
-        }
-    }
-
-    #[test]
-    fn test_decoding_err_unsigned() {
-        let actual: Result<MinecraftUnsignedByte, io::Error> = VecDeque::from(vec![]).decode();
-
-        assert!(actual.is_err());
-    }
-
-    #[test]
-    fn test_encoding_unsigned() {
-        let mappings: Vec<(MinecraftUnsignedByte, Vec<u8>)> =
-            vec![(0x01, vec![0x01]), (0x00, vec![0x00])];
-
-        for mapping in mappings {
-            assert_eq!(VecDeque::from(mapping.1), mapping.0.encode());
-        }
-    }
-
-    #[test]
-    fn test_decoding_signed() {
+    fn test_decoding() {
         let mappings: Vec<(MinecraftByte, Vec<u8>)> = vec![
                 (0x01, vec![0x01]),
                 (0x00, vec![0x00]),
@@ -107,14 +52,14 @@ mod tests {
     }
 
     #[test]
-    fn test_decoding_err_signed() {
+    fn test_decoding_err() {
         let actual: Result<MinecraftByte, io::Error> = VecDeque::from(vec![]).decode();
 
         assert!(actual.is_err());
     }
 
     #[test]
-    fn test_encoding_signed() {
+    fn test_encoding() {
         let mappings: Vec<(MinecraftByte, Vec<u8>)> = vec![
                 (0x01, vec![0x01]),
                 (0x00, vec![0x00]),
